Convert DoublePinchMode to enum
* Added automatic mode to provide default behavior where doubleTapGesture is enabled by default when watchOS is 11.0 or later, and watch is series 9 or later.
